Static json files issue
Didn't know the json's would be static, so
the whole point of copying the json's to dist
serves not much purpose.
Still implemented this way. Didn't make a seperate
commit for that changes and that bites me in the
*ss.
This commit contains changes to taskChange, so
now we can add new tasks to the list. But as noted
this is non persistant data.